% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfc.R
\name{coef.spfc}
\alias{coef.spfc}
\title{Extract SPFC Fit from Complete Workflow}
\usage{
\method{coef}{spfc}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc()}.}

\item{...}{Additional arguments.}
}
\value{
Estimated SPFC direction matrix from the fitted SPFC component.
}
\description{
Extract SPFC Fit from Complete Workflow
}
