% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{coef.spfc_fit}
\alias{coef.spfc_fit}
\title{Extract SPFC Coefficients}
\usage{
\method{coef}{spfc_fit}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc_fit()}.}

\item{...}{Additional arguments.}
}
\value{
Numeric matrix of estimated SPFC directions.
}
\description{
Extracts the estimated SPFC direction matrix.
}
