% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_tools.R
\name{create_train_test_split}
\alias{create_train_test_split}
\title{Create Train-test Split}
\usage{
create_train_test_split(y, test_fraction = 0.3, seed = 123, stratify = TRUE)
}
\arguments{
\item{y}{Response vector.}

\item{test_fraction}{Proportion assigned to test set.}

\item{seed}{Random seed.}

\item{stratify}{Logical. If TRUE, stratify categorical responses.}
}
\value{
A list with train and test indices.
}
\description{
Create Train-test Split
}
\examples{
y <- seq_len(20)
create_train_test_split(
  y, test_fraction = 0.25, seed = 123, stratify = FALSE
)
}
