% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfc.R
\name{fitted.spfc}
\alias{fitted.spfc}
\title{Extract Reduced Scores from Complete Workflow}
\usage{
\method{fitted}{spfc}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc()}.}

\item{...}{Additional arguments.}
}
\value{
Reduced SPFC score matrix.
}
\description{
Extract Reduced Scores from Complete Workflow
}
