% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{fitted.spfc_fit}
\alias{fitted.spfc_fit}
\title{Fitted SPFC Scores}
\usage{
\method{fitted}{spfc_fit}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc_fit()}.}

\item{...}{Additional arguments.}
}
\value{
Numeric matrix of reduced SPFC scores.
}
\description{
Extracts the reduced SPFC score matrix.
}
