% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{plot_dimension_selection}
\alias{plot_dimension_selection}
\title{Plot Dimension-Selection Criteria}
\usage{
plot_dimension_selection(
  dsel,
  criterion = c("AIC", "BIC", "CAIC", "ICOMP_IFIM", "ICOMP_MISSPEC", "CICOMP")
)
}
\arguments{
\item{dsel}{Object returned by \code{spfc_select_dimension()}.}

\item{criterion}{Character. Criterion to plot.}
}
\value{
Invisibly returns the plotted data.
}
\description{
Plots information criteria returned by \code{spfc_select_dimension()}.
}
\examples{
set.seed(1)
X <- matrix(rnorm(180), nrow = 30, ncol = 6)
y <- X[, 1] - 0.7 * X[, 2] + rnorm(30, sd = 0.5)
dsel <- spfc_select_dimension(
  X, y, d_grid = 1:2, cov_method = 'oas',
  ytype = 'continuous', poly_degree = 2, verbose = FALSE
)
plot_dimension_selection(dsel, criterion = 'ICOMP_IFIM')
}
