% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_logistic.R
\name{predict.ridge_logistic_fit}
\alias{predict.ridge_logistic_fit}
\title{Predict from Ridge Logistic Regression}
\usage{
\method{predict}{ridge_logistic_fit}(object, newdata, type = c("class", "response"), ...)
}
\arguments{
\item{object}{Object returned by \code{fit_ridge_logistic()}.}

\item{newdata}{Numeric reduced predictor matrix.}

\item{type}{Character. One of \code{"class"} or \code{"response"}.}

\item{...}{Additional arguments.}
}
\value{
Predicted classes or probabilities.
}
\description{
Predict from Ridge Logistic Regression
}
