% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfc.R
\name{predict.spfc}
\alias{predict.spfc}
\title{Predict SPFC Scores from Complete Workflow}
\usage{
\method{predict}{spfc}(object, newdata, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc()}.}

\item{newdata}{Numeric matrix or data frame of new predictor observations.}

\item{...}{Additional arguments.}
}
\value{
Numeric matrix of predicted SPFC scores.
}
\description{
Predict SPFC Scores from Complete Workflow
}
