% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{predict.spfc_fit}
\alias{predict.spfc_fit}
\title{Predict SPFC Scores for New Data}
\usage{
\method{predict}{spfc_fit}(object, newdata, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc_fit()}.}

\item{newdata}{Numeric matrix or data frame of new predictor observations.}

\item{...}{Additional arguments.}
}
\value{
Numeric matrix of predicted SPFC scores.
}
\description{
Projects new predictor observations into the estimated SPFC reduced space.
}
