% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfc.R
\name{print.spfc}
\alias{print.spfc}
\title{Print Complete SPFC Workflow}
\usage{
\method{print}{spfc}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{spfc()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Complete SPFC Workflow
}
