% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_spfc.R
\name{print.spfc_benchmark}
\alias{print.spfc_benchmark}
\title{Print SPFC Benchmark}
\usage{
\method{print}{spfc_benchmark}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{benchmark_spfc()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print SPFC Benchmark
}
