% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.spfc_fit}
\alias{print.spfc_fit}
\title{Print an SPFC Fit}
\usage{
\method{print}{spfc_fit}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{spfc_fit()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an SPFC Fit
}
