% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfc.R
\name{print.summary.spfc}
\alias{print.summary.spfc}
\title{Print Summary of Complete SPFC Workflow}
\usage{
\method{print}{summary.spfc}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{summary.spfc()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Summary of Complete SPFC Workflow
}
