% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_spfc.R
\name{print.summary.spfc_benchmark}
\alias{print.summary.spfc_benchmark}
\title{Print Summary of SPFC Benchmark}
\usage{
\method{print}{summary.spfc_benchmark}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{summary.spfc_benchmark()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Summary of SPFC Benchmark
}
