% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.summary.spfc_fit}
\alias{print.summary.spfc_fit}
\title{Print Summary of an SPFC Fit}
\usage{
\method{print}{summary.spfc_fit}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{summary.spfc_fit()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Summary of an SPFC Fit
}
