% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{run_spfc_simulation}
\alias{run_spfc_simulation}
\title{Run a Small SPFC Simulation Study}
\usage{
run_spfc_simulation(
  response_type = c("continuous", "categorical"),
  nrep = 10,
  n = 100,
  p = 50,
  d = 1,
  s = 5,
  rho_x = 0.5,
  snr = 1,
  signal_strength = 1,
  cov_methods = c("mec", "oas", "sre", "sde", "cse"),
  variable_methods = c("adaptive_weighted_l1", "c1f_extension"),
  seed = 123
)
}
\arguments{
\item{response_type}{Character. \code{"continuous"} or \code{"categorical"}.}

\item{nrep}{Number of Monte Carlo replicates.}

\item{n}{Sample size.}

\item{p}{Number of predictors.}

\item{d}{Structural dimension.}

\item{s}{Number of active variables.}

\item{rho_x}{Predictor correlation.}

\item{snr}{Signal-to-noise ratio for continuous response.}

\item{signal_strength}{Signal strength for categorical response.}

\item{cov_methods}{Covariance estimators to compare.}

\item{variable_methods}{Variable-selection methods to compare.}

\item{seed}{Random seed.}
}
\value{
A data frame of simulation results.
}
\description{
Run a Small SPFC Simulation Study
}
\examples{
\donttest{
ans <- run_spfc_simulation(
  response_type = 'continuous', nrep = 1, n = 30, p = 8,
  d = 1, s = 3, rho_x = 0.3, snr = 2,
  cov_methods = c('oas', 'sre'),
  variable_methods = 'adaptive_weighted_l1', seed = 123
)
ans[, c('cov_method', 'rmse', 'subspace_distance')]
}
}
