% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfc.R
\name{summary.spfc}
\alias{summary.spfc}
\title{Summarise Complete SPFC Workflow}
\usage{
\method{summary}{spfc}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc()}.}

\item{...}{Additional arguments.}
}
\value{
A list of class \code{summary.spfc}.
}
\description{
Summarise Complete SPFC Workflow
}
