% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_spfc.R
\name{summary.spfc_benchmark}
\alias{summary.spfc_benchmark}
\title{Summarise SPFC Benchmark}
\usage{
\method{summary}{spfc_benchmark}(object, metric = NULL, smaller_is_better = NULL, ...)
}
\arguments{
\item{object}{Object returned by \code{benchmark_spfc()}.}

\item{metric}{Optional selection metric.}

\item{smaller_is_better}{Optional logical indicating whether smaller values
are better.}

\item{...}{Additional arguments.}
}
\value{
A list of class \code{summary.spfc_benchmark}.
}
\description{
Summarise SPFC Benchmark
}
