% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.spfc_fit}
\alias{summary.spfc_fit}
\title{Summarise an SPFC Fit}
\usage{
\method{summary}{spfc_fit}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{spfc_fit()}.}

\item{...}{Additional arguments.}
}
\value{
A list containing key SPFC summary components.
}
\description{
Summarise an SPFC Fit
}
